block: rename tests file to test_helpers
The tests.rs creates a module named `tests`, which collides with the
rust convention to name test modules `test`.
Rename the file to test_helpers.rs to prevent the collision before
renaming `unit_tests` modules to the Rust idiomatic `tests`.

On-behalf-of: SAP julian.schindel@sap.com
